+++ src/query/term_query/mod.rs
@@ -9,12 +9,12 @@ pub use self::term_weight::TermWeight;
 #[cfg(test)]
 mod tests {
 
-    use crate::assert_nearly_equals;
     use crate::collector::TopDocs;
     use crate::docset::DocSet;
     use crate::postings::compression::COMPRESSION_BLOCK_SIZE;
     use crate::query::{Query, QueryParser, Scorer, TermQuery};
     use crate::schema::{Field, IndexRecordOption, Schema, STRING, TEXT};
+    use crate::{assert_nearly_equals, DocAddress};
     use crate::{Index, Term, TERMINATED};
 
     #[test]
@@ -178,5 +178,41 @@ mod tests {
             format!("{:?}", term_query),
             "TermQuery(Term(field=1,bytes=[104, 101, 108, 108, 111]))"
         );
+    }
+
+    #[test]
+    fn test_term_query_explain() -> crate::Result<()> {
+        let mut schema_builder = Schema::builder();
+        let text_field = schema_builder.add_text_field("text", TEXT);
+        let schema = schema_builder.build();
+        let index = Index::create_in_ram(schema);
+        let mut index_writer = index.writer_for_tests().unwrap();
+        index_writer.add_document(doc!(text_field=>"b"));
+        index_writer.add_document(doc!(text_field=>"a"));
+        index_writer.add_document(doc!(text_field=>"a"));
+        index_writer.add_document(doc!(text_field=>"b"));
+        index_writer.commit()?;
+        let term_a = Term::from_field_text(text_field, "a");
+        let term_query = TermQuery::new(term_a, IndexRecordOption::Basic);
+        let searcher = index.reader()?.searcher();
+        {
+            let explanation = term_query.explain(&searcher, DocAddress(0u32, 1u32))?;
+            assert_nearly_equals!(explanation.value(), 0.6931472f32);
+        }
+        {
+            let explanation_err = term_query.explain(&searcher, DocAddress(0u32, 0u32));
+            assert!(matches!(
+                explanation_err,
+                Err(crate::TantivyError::InvalidArgument(_msg))
+            ));
+        }
+        {
+            let explanation_err = term_query.explain(&searcher, DocAddress(0u32, 3u32));
+            assert!(matches!(
+                explanation_err,
+                Err(crate::TantivyError::InvalidArgument(_msg))
+            ));
+        }
+        Ok(())
     }
 }